+++ sitepulse-engine-app/src/main/java/com/sitepulse/engine/metrics/domain/port/WeeklyMetricCatalogRepository.java
@@ -1,6 +1,7 @@
 package com.sitepulse.engine.metrics.domain.port;
 
 import com.sitepulse.engine.metrics.domain.model.WeeklyMetric;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
@@ -13,5 +14,5 @@ public interface WeeklyMetricCatalogRepository {
 
     List<WeeklyMetric> findLatest(Integer projectId, int limit);
 
-    Double findAverageActivityBefore(Integer projectId, LocalDate weekStart);
+    BigDecimal findAverageActivityBefore(Integer projectId, LocalDate weekStart);
 }
